medical_cabinet/backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
aniso8601==9.0.1
attrs==24.3.0
bcrypt==4.2.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
dnspython==2.2.1
ecdsa==0.19.0
Flask==2.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.2
propcache==0.2.1
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.10.1
pymongo==4.2.0
pyperclip==1.9.0
python-dotenv==1.0.1
python-jose==3.3.0
python-keycloak==0.23.0
pytz==2024.2
requests==2.32.3
rsa==4.9
six==1.17.0
twilio==9.4.1
urllib3==1.26.20
Werkzeug==2.2.2
yarl==1.18.3
